requirements.txt
accelerate
apple-foundation-models
einops
pillow
pyexiftool
pytesseract
pyvips>=2.2
torch
transformers
